Pathing Refactor + Review Comments
