+++ App.config
@@ -33,6 +33,8 @@
         <add key="HiddenAreas" value="Rogue Encampment,Lut Gholein,Kurast Docktown,The Pandemonium Fortress,Harrogath,Forgotten Tower,Nihlathaks Temple" />
 
 		<add key="MovementMode" value="Walking" />	<!-- "Teleport" or "Walking" -->
+		<add key="TpRange" value="45" />
+		<add key="PathUpdateTime" value="500" />
 		
 		<add key="MonsterColor.Normal" value="DarkRed"/>
 		<add key="MonsterColor.Elite" value="DarkOrange"/>
@@ -57,14 +59,18 @@
         <add key="NextArea.IconShape" value="Rectangle" />
         <add key="NextArea.IconSize" value="10" />
         <add key="NextArea.LabelColor" value="Chartreuse" />
-		<add key="NextArea.PathColor" value="Chartreuse" />
+		<add key="NextArea.LineColor" value="Chartreuse" />
+		<add key="NextArea.ArrowHeadSize" value="10" />
+		<!-- <add key="NextArea.PathColor" value="Chartreuse" /> -->
 
         <add key="PreviousArea.IconColor" value="255, 0, 149" />
         <add key="PreviousArea.IconShape" value="Rectangle" />
         <add key="PreviousArea.IconSize" value="10" />
         <add key="PreviousArea.LabelColor" value="Chartreuse" />
 
-		<add key="Quest.PathColor" value="orange" />
+		<add key="Quest.LineColor" value="Chartreuse" />
+		<add key="Quest.ArrowHeadSize" value="5" />
+		<!-- <add key="Quest.PathColor" value="orange" /> -->
 
         <add key="MapColor[0]" value="50, 50, 50" />
         <add key="MapColor[2]" value="10, 51, 23" />
